@@ -1,5 +1,6 @@
 {
 	"name": "intelliskript",
+	"displayName": "IntelliSkript",
 	"description": "IntelliSense for skript",
 	"author": "John Heikens",
 	"license": "GNU",
@@ -12,22 +13,29 @@
 	"publisher": "JohnHeikens",
 	"categories": [
 		"Programming Languages",
-		"Linters"
+		"Linters",
+		"Formatters"
 	],
 	"keywords": [
 		"skript",
 		"intelliskript",
-		"intellisense"
+		"intellisense",
+		"format",
+		"multi-root ready"
 	],
 	"engines": {
-		"vscode": "^1.63.0"
+		"vscode": "^1.85.0"
 	},
 	"activationEvents": [
 		"onLanguage:Sk-VSC",
 		"onLanguage:sk",
 		"onLanguage:skriptlang"
 	],
-	"main": "./client/out/extension",
+	"extensionKind": [
+		"ui",
+		"workspace"
+	],
+	"browser": "./client/dist/browserClientMain",
 	"contributes": {
 		"languages": [
 			{
@@ -50,57 +58,64 @@
 				}
 			}
 		],
-		"configuration": {
-			"type": "object",
-			"title": "IntelliSkript",
-			"properties": {
-				"IntelliSkript.ErrorExempts": {
-					"scope": "resource",
-					"type": "string",
-					"default": "",
-					"description": "A RegExp which allows you to filter out annoying errors"
-				},
-				"IntelliSkript.UseColorTheme": {
-					"scope": "resource",
-					"type": "boolean",
-					"default": true,
-					"description": "Disable this and reload VS code if you have a skript color theme plugin installed"
-				},
-				"intelliSkript.trace.server": {
-					"scope": "window",
-					"type": "string",
-					"enum": [
-						"off",
-						"messages",
-						"verbose"
-					],
-					"default": "off",
-					"description": "Traces the communication between VS Code and the language server."
+		"configuration": [
+			{
+				"type": "object",
+				"title": "IntelliSkript",
+				"id": "IntelliSkript",
+				"properties": {
+					"IntelliSkript.ErrorExempts": {
+						"scope": "resource",
+						"type": "string",
+						"default": "",
+						"description": "A RegExp which allows you to filter out annoying errors"
+					},
+					"IntelliSkript.UseColorTheme": {
+						"scope": "resource",
+						"type": "boolean",
+						"default": true,
+						"description": "Disable this and reload VS code if you have a skript color theme plugin installed"
+					},
+					"intelliSkript.trace.server": {
+						"scope": "window",
+						"type": "string",
+						"enum": [
+							"off",
+							"messages",
+							"verbose"
+						],
+						"default": "off",
+						"description": "Traces the communication between VS Code and the language server."
+					}
 				}
 			}
-		}
+		]
 	},
 	"scripts": {
-		"vscode:prepublish": "webpack --mode production",
-		"webpack": "webpack --mode development",
-		"webpack-dev": "webpack --mode development --watch",
-		"compile": "npm run webpack",
-		"watch": "tsc -watch -p ./",
-		"postinstall": "node ./node_modules/vscode/bin/install"
+		"vscode:prepublish": "npm run compile",
+		"compile": "webpack",
+		"watch": "webpack --watch",
+		"package": "webpack --mode production --devtool hidden-source-map",
+		"lint": "eslint ./client/src ./server/src --ext .ts,.tsx",
+		"postinstall": "cd client && npm install && cd ../server && npm install && cd ..",
+		"chrome": "npm run compile && vscode-test-web --browserType=chromium --extensionDevelopmentPath=. ./test-data"
 	},
 	"devDependencies": {
-		"@types/mocha": "^9.1.0",
-		"@types/node": "^16.11.7",
-		"@typescript-eslint/eslint-plugin": "^5.42.0",
-		"@typescript-eslint/parser": "^5.42.0",
-		"eslint": "^8.26.0",
-		"mocha": "^9.2.1",
+		"@typescript-eslint/eslint-plugin": "^6.20.0",
+		"@typescript-eslint/parser": "^6.20.0",
+		"@vscode/test-web": "^0.0.51",
+		"copy-webpack-plugin": "^12.0.2",
+		"eslint": "^8.56.0",
+		"path-browserify": "^1.0.1",
+		"raw-loader": "^4.0.2",
 		"ts-loader": "^9.5.1",
-		"typescript": "^4.9.5",
-		"webpack": "^5.93.0",
+		"ts-node": "^10.9.2",
+		"ts-webpack": "^1.0.0",
+		"typescript": "^5.5.4",
 		"webpack-cli": "^5.1.4"
 	},
 	"dependencies": {
-		"vscode-uri": "^3.0.7"
+		"buffer": "^6.0.3",
+		"vscode-uri": "^3.0.8"
 	}
 }
