@@ -1,11 +1,5 @@
 
-import * as path from 'path';
-
-export const ServerDirectory = path.resolve(__dirname, "..");
-export const ServerAssetsDirectory = path.join(ServerDirectory, "Assets");
-export const AddonSkFilesDirectory = path.join(ServerAssetsDirectory, "Addons");
-export const IsReleaseMode = ServerDirectory.includes('johnheikens.intelliskript');
-export const ServerSrcDirectory = path.join(ServerDirectory, "src");
+export const IsDebugMode = !__dirname.includes('johnheikens.intelliskript');
 /**
  * capture groups:
  * [0]: returns "-" or ""
@@ -15,4 +9,3 @@ export const ServerSrcDirectory = path.join(ServerDirectory, "src");
  */
 export const NumberRegExp = /(?:(?<=(?: |^))(-){0,1})(\d+)(?:\.(\d+))?(?!\.)\b/;
 export const BooleanRegExp = /(?<=( |^))true|false(?=( |$))/;
-export const skriptFileHeader = "#AUTOMATICALLY GENERATED SKRIPT FILE\n#COPYRIGHT JOHN HEIKENS\n#https://github.com/JohnHeikens/IntelliSkript\n"
