@@ -6,7 +6,7 @@
  * @param compare should return a < b
  * @returns 
  */
-export function sortedIndex<t>(array: Array<t>, value: t, compare: (a: t, b: t) => boolean, low: number = 0, high: number = array.length) {
+export function sortedIndex<tArr, tComp>(array: Array<tArr>, value: tComp, compare: (a: tArr, b: tComp) => boolean, low: number = 0, high: number = array.length) {
     while (low < high) {
         var mid = (low + high) >>> 1;
         //array[mid] < value
@@ -14,4 +14,9 @@ export function sortedIndex<t>(array: Array<t>, value: t, compare: (a: t, b: t) 
         else high = mid;
     }
     return low;
+}
+
+export function insertSorted<t>(array: Array<t>, value: t, compare: (a: t, b: t) => boolean) {
+    const index = sortedIndex(array, value, compare);
+    array.splice(index, 0, value);
 }
