@@ -1,44 +1,121 @@
+/* eslint-disable no-undef */
+/* eslint-disable @typescript-eslint/no-var-requires */
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
 //@ts-check
+'use strict';
 
-(function () {
-  'use strict';
+//@ts-check
+/** @typedef {import('webpack').Configuration} WebpackConfig **/
 
-  const path = require('path');
+const path = require('path');
+const CopyWebpackPlugin = require('copy-webpack-plugin');
 
-  /**@type {import('webpack').Configuration}*/
-  const config = {
-    target: 'node', // vscode extensions run in a Node.js-context 📖 -> https://webpack.js.org/configuration/node/
+/** @type WebpackConfig */
+const browserClientConfig = {
+	context: path.join(__dirname, 'client'),
+	mode: 'none',
+	target: 'webworker', // web extensions run in a webworker context
+	entry: {
+		browserClientMain: './src/browserClientMain.ts',
+	},
+	output: {
+		filename: '[name].js',
+		path: path.join(__dirname, 'client', 'dist'),
+		libraryTarget: 'commonjs',
+		devtoolModuleFilenameTemplate: '../[resource-path]'
+	},
+	resolve: {
+		mainFields: ['module', 'main'],
+		extensions: ['.ts', '.js'], // support ts-files and js-files
+		alias: {},
+		fallback: {
+			path: require.resolve('path-browserify'),
+		},
+	},
+	module: {
+		rules: [
+			{
+				test: /\.ts$/,
+				exclude: /node_modules/,
+				use: [
+					{
+						loader: 'ts-loader',
+					},
+				],
+			},
+		],
+	},
+	externals: {
+		vscode: 'commonjs vscode', // ignored because it doesn't exist
+	},
+	performance: {
+		hints: false,
+	},
+	devtool: 'nosources-source-map',
+};
 
-    entry: './client/src/extension.ts', // the entry point of this extension, 📖 -> https://webpack.js.org/configuration/entry-context/
-    output: {
-      // the bundle is stored in the 'dist' folder (check package.json), 📖 -> https://webpack.js.org/configuration/output/
-      path: path.resolve(__dirname, 'dist'),
-      filename: 'extension.js',
-      clean: true, //clean the dist folder for each time webpack is run
-      libraryTarget: 'commonjs2',
-      devtoolModuleFilenameTemplate: '../[resource-path]'
-    },
-    devtool: 'source-map',
-    externals: {
-      vscode: 'commonjs vscode' // the vscode-module is created on-the-fly and must be excluded. Add other modules that cannot be webpack'ed, 📖 -> https://webpack.js.org/configuration/externals/
-    },
-    resolve: {
-      // support reading TypeScript and JavaScript files, 📖 -> https://github.com/TypeStrong/ts-loader
-      extensions: ['.ts', '.js']
-    },
-    module: {
-      rules: [
-        {
-          test: /\.ts$/,
-          exclude: /node_modules/,
-          use: [
-            {
-              loader: 'ts-loader'
-            }
-          ]
-        }
-      ]
-    }
-  };
-  module.exports = config;
-}());
+/** @type WebpackConfig */
+const browserServerConfig = {
+	context: path.join(__dirname, 'server'),
+	mode: 'none',
+	target: 'webworker', // web extensions run in a webworker context
+	entry: {
+		browserServerMain: './src/browserServerMain.ts',
+	},
+	output: {
+		filename: '[name].js',
+		path: path.join(__dirname, 'server', 'dist'),
+		libraryTarget: 'var',
+		library: 'serverExportVar',
+		devtoolModuleFilenameTemplate: '../[resource-path]',
+		assetModuleFilename: 'assets/[hash][ext][query]'
+	},
+	resolve: {
+		mainFields: ['module', 'main'],
+		extensions: ['.ts', '.js', '.sk'], // support ts-files, js-files and sk files (the addon folder)
+		alias: {},
+		fallback: {
+			//path: require.resolve("path-browserify")
+		},
+	},
+	plugins: [
+		new CopyWebpackPlugin({
+			patterns: [
+				{ from: 'assets/addons', to: 'assets/addons' },
+			],
+		})
+	],
+	module: {
+		rules: [
+			{
+				test: /\.ts$/,
+				exclude: /node_modules/,
+				use: [
+					{
+						loader: 'ts-loader',
+					},
+				],
+			},
+			{
+				test: /\.sk$/,
+				use: 'raw-loader',
+			},
+		],
+	},
+	externals: {
+		vscode: 'commonjs vscode', // ignored because it doesn't exist
+	},
+	performance: {
+		hints: false,
+	},
+	devtool: 'nosources-source-map',
+};
+
+const webpack = require('webpack');
+
+
+module.exports = [browserClientConfig, browserServerConfig];
